feat(quiz): sort the 题库 by last-activity time (status_update_time)
Add a topic-style status_update_time to galgame_quiz: defaults to the creation
time, bumped when a user answers the quiz or the author edits it. The
/galgame-quiz list gains a "最近活跃" sort (now the default) and shows this time on
each card.

- Migration 052: adds status_update_time (backfilled to created, default now(),
  indexed); sort field "update_time" → q.status_update_time.
- Bump on answer (BumpAnswerStats) + author edit (UpdateQuiz fields); seeded to
  now() on create.
- FE: "最近活跃" sort option (now default), card shows status_update_time, +type/const.

Also fixes a latent bug in migration 051's feed trigger: galgame_quiz.id is BIGINT
(other feed sources are INT), so the uncast NEW.id/OLD.id failed to resolve
feed_upsert/feed_delete — dormant during 051's direct-INSERT backfill, but it would
have errored on the first quiz answer/edit (and on 052's UPDATE). Cast to integer
in 051 (source of truth) + re-assert the fixed function in 052 (repairs any DB that
already ran the broken 051).

+++ apps/api/migrations/051_feed_galgame_quiz.up.sql
@@ -5,10 +5,13 @@
 -- it is never NSFW-dropped by the feed's galgame enrichment. content = the 题干.
 -- Idempotent (reuses the shared feed_upsert / feed_delete helpers).
 
+-- NOTE galgame_quiz.id is BIGINT (unlike the other feed sources' INT ids), so
+-- NEW.id / OLD.id must be cast to integer or the feed_upsert/feed_delete calls
+-- (INT source_id params) fail to resolve ("function does not exist").
 CREATE OR REPLACE FUNCTION feed_sync_galgame_quiz() RETURNS trigger AS $$
 BEGIN
-    IF TG_OP = 'DELETE' THEN PERFORM feed_delete('GALGAME_QUIZ_CREATION', OLD.id); RETURN OLD; END IF;
-    PERFORM feed_upsert('GALGAME_QUIZ_CREATION', NEW.id, NEW.user_id, 0, NEW.question, '/galgame-quiz/' || NEW.id, false, NEW.created);
+    IF TG_OP = 'DELETE' THEN PERFORM feed_delete('GALGAME_QUIZ_CREATION', OLD.id::integer); RETURN OLD; END IF;
+    PERFORM feed_upsert('GALGAME_QUIZ_CREATION', NEW.id::integer, NEW.user_id, 0, NEW.question, '/galgame-quiz/' || NEW.id, false, NEW.created);
     RETURN NEW;
 END $$ LANGUAGE plpgsql;
